+++ graphql/src/graphql_intf.ml
@@ -36,48 +36,48 @@ module type Schema = sig
             ('ctx, 'src option) typ
 
   module Arg : sig
-    type (_, _) arg
-    type (_, _) arg_typ
+    type _ arg
+    type _ arg_typ
 
     type (_, _) arg_list =
       | [] : ('a, 'a) arg_list
-      | (::) : ('b, 'c -> 'b) arg * ('a, 'b) arg_list -> ('a, 'c -> 'b) arg_list
+      | (::) : 'a arg * ('b, 'c) arg_list -> ('b, 'a -> 'c) arg_list
 
     val arg : ?doc:string ->
               string ->
-              typ:('a, 'b -> 'a) arg_typ ->
-              ('a, 'b -> 'a) arg
+              typ:'a arg_typ ->
+              'a arg
 
     val arg' : ?doc:string ->
                string ->
-               typ:('a, 'b option -> 'a) arg_typ ->
-               default:'b ->
-               ('a, 'b -> 'a) arg
+               typ:'a option arg_typ ->
+               default:'a ->
+               'a arg
 
     val scalar : ?doc:string ->
                  string ->
-                 coerce:(Graphql_parser.const_value -> ('b, string) result) ->
-                 ('a, 'b option -> 'a) arg_typ
+                 coerce:(Graphql_parser.const_value -> ('a, string) result) ->
+                 'a option arg_typ
 
     val enum : ?doc:string ->
                string ->
-               values:'b enum_value list ->
-               ('a, 'b option -> 'a) arg_typ
+               values:'a enum_value list ->
+               'a option arg_typ
 
     val obj : ?doc:string ->
               string ->
-              fields:('c, 'b) arg_list ->
+              fields:('a, 'b) arg_list ->
               coerce:'b ->
-              ('a, 'c option -> 'a) arg_typ
+              'a option arg_typ
 
     (* Argument constructors *)
-    val int : ('a, int option -> 'a) arg_typ
-    val string : ('a, string option -> 'a) arg_typ
-    val bool : ('a, bool option -> 'a) arg_typ
-    val float : ('a, float option -> 'a) arg_typ
-    val guid : ('a, string option -> 'a) arg_typ
-    val list : ('a, 'b -> 'a) arg_typ -> ('a, 'b list option -> 'a) arg_typ
-    val non_null : ('a, 'b option -> 'a) arg_typ -> ('a, 'b -> 'a) arg_typ
+    val int : int option arg_typ
+    val string : string option arg_typ
+    val bool : bool option arg_typ
+    val float : float option arg_typ
+    val guid : string option arg_typ
+    val list : 'a arg_typ -> 'a list option arg_typ
+    val non_null : 'a option arg_typ -> 'a arg_typ
   end
 
   type 'ctx resolve_params = {