+++ graphql/test/introspection_test.ml
@@ -1,5 +1,7 @@
 open Graphql
 
+let test_query schema query = Test_common.test_query schema () query
+
 let suite = [
   ("not deprecated", `Quick, fun () ->
     let schema = Schema.(schema [
@@ -9,7 +11,21 @@ let suite = [
         ~args:Arg.[]
         ~resolve:(fun _ _ -> Some "")
     ]) in
-    Test_common.test_query schema () "{ __schema { queryType { fields { isDeprecated deprecationReason } } } }" "{\"data\":{\"__schema\":{\"queryType\":{\"fields\":[{\"isDeprecated\":false,\"deprecationReason\":null}]}}}}"
+    let query = "{ __schema { queryType { fields { isDeprecated deprecationReason } } } }" in
+    test_query schema query (`Assoc [
+      "data", `Assoc [
+        "__schema", `Assoc [
+          "queryType", `Assoc [
+            "fields", `List [
+              `Assoc [
+                "isDeprecated", `Bool false;
+                "deprecationReason", `Null
+              ]
+            ]
+          ]
+        ]
+      ]
+    ])
   );
   ("default deprecation", `Quick, fun () ->
     let schema = Schema.(schema [
@@ -18,7 +34,21 @@ let suite = [
         ~args:Arg.[]
         ~resolve:(fun _ _ -> Some "")
     ]) in
-    Test_common.test_query schema () "{ __schema { queryType { fields { isDeprecated deprecationReason } } } }" "{\"data\":{\"__schema\":{\"queryType\":{\"fields\":[{\"isDeprecated\":false,\"deprecationReason\":null}]}}}}"
+    let query = "{ __schema { queryType { fields { isDeprecated deprecationReason } } } }" in
+    test_query schema query (`Assoc [
+      "data", `Assoc [
+        "__schema", `Assoc [
+          "queryType", `Assoc [
+            "fields", `List [
+              `Assoc [
+                "isDeprecated", `Bool false;
+                "deprecationReason", `Null
+              ]
+            ]
+          ]
+        ]
+      ]
+    ])
   );
   ("deprecated-without-reason", `Quick, fun () ->
     let schema = Schema.(schema [
@@ -28,7 +58,21 @@ let suite = [
         ~args:Arg.[]
         ~resolve:(fun _ _ -> Some "")
     ]) in
-    Test_common.test_query schema () "{ __schema { queryType { fields { isDeprecated deprecationReason } } } }" "{\"data\":{\"__schema\":{\"queryType\":{\"fields\":[{\"isDeprecated\":true,\"deprecationReason\":null}]}}}}"
+    let query = "{ __schema { queryType { fields { isDeprecated deprecationReason } } } }" in
+    test_query schema query (`Assoc [
+      "data", `Assoc [
+        "__schema", `Assoc [
+          "queryType", `Assoc [
+            "fields", `List [
+              `Assoc [
+                "isDeprecated", `Bool true;
+                "deprecationReason", `Null
+              ]
+            ]
+          ]
+        ]
+      ]
+    ])
   );
   ("deprecated with reason", `Quick, fun () ->
     let schema = Schema.(schema [
@@ -38,6 +82,46 @@ let suite = [
         ~args:Arg.[]
         ~resolve:(fun _ _ -> Some "")
     ]) in
-    Test_common.test_query schema () "{ __schema { queryType { fields { isDeprecated deprecationReason } } } }" "{\"data\":{\"__schema\":{\"queryType\":{\"fields\":[{\"isDeprecated\":true,\"deprecationReason\":\"deprecation reason\"}]}}}}"
+    let query = "{ __schema { queryType { fields { isDeprecated deprecationReason } } } }" in
+    test_query schema query (`Assoc [
+      "data", `Assoc [
+        "__schema", `Assoc [
+          "queryType", `Assoc [
+            "fields", `List [
+              `Assoc [
+                "isDeprecated", `Bool true;
+                "deprecationReason", `String "deprecation reason"
+              ]
+            ]
+          ]
+        ]
+      ]
+    ])
+  );
+  ("deduplicates argument types", `Quick, fun () ->
+    let schema = Schema.(schema [
+      field "sum"
+        ~typ:(non_null int)
+        ~args:Arg.[
+          arg "x" ~typ:(non_null int);
+          arg "y" ~typ:(non_null int);
+        ]
+        ~resolve:(fun _ _ x y -> x + y)
+    ]) in
+    let query = "{ __schema { types { name } } }" in
+    test_query schema query (`Assoc [
+      "data", `Assoc [
+        "__schema", `Assoc [
+          "types", `List [
+            `Assoc [
+              "name", `String "Int"
+            ];
+            `Assoc [
+              "name", `String "query"
+            ]
+          ]
+        ]
+      ]
+    ])
   );
 ]