+++ graphql_parser/test/test.ml
@@ -1,0 +1,4 @@
+let () =
+  Alcotest.run "graphql-server" [
+    "parser", Parser_test.suite;
+  ]