+++ CHANGES.md
@@ -1,3 +1,28 @@
+Unreleased
+---------------------------------
+
+- Prevent fragment cycles (#78)
+
+0.5.0 2018-01-18
+---------------------------------
+
+- Depend on angstrom 0.7.0 (#64)
+- Fix parsing of quoted strings (#64)
+- Make custom argument types generalizable (#72)
+- Deduplicate arg types in introspection result (#75)
+- 4.06 compatibility (#76)
+
+0.4.0 2017-09-17
+---------------------------------
+
+- Parse tabs as whitespace (#62)
+- Move parser to separate package (#63)
+
+0.3.0 2017-08-31
+---------------------------------
+
+- Built-in HTTP server for graphql-lwt (#60)
+
 0.2.0 2017-07-02
 ---------------------------------